Created most of the home page
I created the most basic, necessary elements needed for the home page / hero page. It's taken a lot of from the inspiration (Wix site).

Most of the information is complete for this first page. just need to finish up the footer info.

The navbar is still the same.
I added more div's, containers, and a footer at the bottom with a line divider above.