Started the portfolio page
I started working on the Portfolio page.
This contains a grid of projects that client has worked on.
The website viewer will be able to click on a tile on the grid, then a pop-up (or modal in this case) will show, expand to show more information and a bigger picture.

I'm thinking of using modal, because it seems simplest rn. However it doesn't look that good. But it still works :P