Fixed the css import
I had the wrong import href="" for importing the css, for the pages inside the page folder xd

But it's fixed now, and the pages are looking a lot better!

The portfolio page especially, since there are fade in animations for the modal, scale-transform changes for when hovering over a project, and the modal layout is improved.