Fixed text exceeding view width
It's a bit late to notice, but I saw that the "Welcome to my..." text on the homepage actually went pass the device width and container on mobile viewport.
I fixed that by adding a snippet of  CSS with media queries. On smaller viewport, the font shrinks

And I updated the podcast section by contact page. Embedding a podcast interface (sehr cool) and keeping the link